layout/style/test/test_constructed_stylesheet.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test CSSStyleSheet constructor</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css">
</head>
<body>
<pre id="test"></pre>
<script type="application/javascript">
"use strict";

let sheet;

try {
  sheet = new CSSStyleSheet();
  ok(true, "CSSStyleSheet constructor should not throw");
} catch (e) {
  ok(false, "CSSStyleSheet constructor should not throw: " + e);
}

if (sheet) {
  ok(sheet instanceof CSSStyleSheet, "Constructor should create a CSSStyleSheet");
  is(sheet.type, "text/css", "Constructed sheet should have CSS type");
  is(sheet.href, null, "Constructed sheet should not have an href");
  is(sheet.cssRules.length, 0, "Constructed sheet should start with no rules");

  is(sheet.insertRule("#test { color: rgb(1, 2, 3); }", 0), 0,
     "insertRule should work on a constructed sheet");
  is(sheet.cssRules.length, 1, "insertRule should append a rule");
  is(sheet.cssRules[0].selectorText, "#test", "Inserted rule should be readable");

  sheet.deleteRule(0);
  is(sheet.cssRules.length, 0, "deleteRule should remove the rule");

  is(sheet.replaceSync("#test { color: rgb(4, 5, 6); } p { margin-top: 1px; }"),
     undefined, "replaceSync should return undefined");
  is(sheet.cssRules.length, 2, "replaceSync should replace with parsed rules");
  is(sheet.cssRules[0].selectorText, "#test",
     "replaceSync should expose the first parsed rule");
  is(sheet.cssRules[1].selectorText, "p",
     "replaceSync should expose subsequent parsed rules");

  sheet.replaceSync("@import url('neverending_stylesheet_load.sjs'); #test { color: rgb(7, 8, 9); }");
  is(sheet.cssRules.length, 1, "replaceSync should remove @import rules");
  is(sheet.cssRules[0].selectorText, "#test",
     "replaceSync should keep non-import rules after removed imports");

  sheet.replaceSync("}");
  is(sheet.cssRules.length, 0,
     "replaceSync should replace with an empty rule list when nothing parses");

  try {
    document.styleSheets[0].replaceSync("#test { color: rgb(10, 11, 12); }");
    ok(false, "replaceSync should throw for non-constructed sheets");
  } catch (e) {
    is(e.name, "NotAllowedError",
       "replaceSync should reject non-constructed sheets with NotAllowedError");
  }
}
</script>
</body>
</html>
